refactor: better file structure - reduce the number of files in root folder
