+++ lib/dynamic.js
@@ -81,7 +81,7 @@ module.exports = function (fastify, opts, next) {
     let pkg
 
     try {
-      pkg = JSON.parse(fs.readFileSync(path.join(__dirname, 'package.json')))
+      pkg = JSON.parse(fs.readFileSync(path.join(__dirname, '..', 'package.json')))
     } catch (err) {
       return next(err)
     }
