+++ lib/routes.js
@@ -51,13 +51,13 @@ function fastifySwagger (fastify, opts, next) {
 
   // serve swagger-ui with the help of fastify-static
   fastify.register(fastifyStatic, {
-    root: path.join(__dirname, 'static'),
+    root: path.join(__dirname, '..', 'static'),
     prefix: staticPrefix,
     decorateReply: false
   })
 
   fastify.register(fastifyStatic, {
-    root: opts.baseDir || __dirname,
+    root: opts.baseDir || path.join(__dirname, '..'),
     serve: false
   })
