+++ lib/static.js
