+++ test/static.js
@@ -5,7 +5,7 @@ const t = require('tap')
 const test = t.test
 const Fastify = require('fastify')
 const fastifySwagger = require('../index')
-const fastifySwaggerDynamic = require('../dynamic')
+const fastifySwaggerDynamic = require('../lib/dynamic')
 const yaml = require('js-yaml')
 
 const resolve = require('path').resolve
@@ -615,7 +615,7 @@ test('inserts default package name', t => {
   const testPackageJSON = path.join(__dirname, '../examples/test-package.json')
 
   path.join = (...args) => {
-    if (args[1] === 'package.json') {
+    if (args[2] === 'package.json') {
       return testPackageJSON
     }
     return originalPathJoin(...args)
@@ -650,7 +650,7 @@ test('throws an error if cannot parse package\'s JSON', t => {
   const testPackageJSON = path.join(__dirname, '')
 
   path.join = (...args) => {
-    if (args[1] === 'package.json') {
+    if (args[2] === 'package.json') {
       return testPackageJSON
     }
     return originalPathJoin(...args)
